chore: community 관련 패키지 구조 변경 (#13)
