+++ api-module/src/main/java/org/devridge/api/domain/community/dto/request/CreateCommunityRequest.java
@@ -1,4 +1,4 @@
-package org.devridge.api.domain.community;
+package org.devridge.api.domain.community.dto.request;
 
 import lombok.Getter;
