+++ api-module/src/main/java/org/devridge/api/domain/community/entity/CommunityScrap.java
@@ -1,11 +1,9 @@
-package org.devridge.api.domain.communitycommentlikedislike;
+package org.devridge.api.domain.community.entity;
 
 import java.time.LocalDateTime;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -15,7 +13,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.devridge.api.domain.communitycomment.CommunityComment;
+import org.devridge.api.domain.community.entity.id.CommunityScrapId;
 import org.devridge.api.domain.member.entity.Member;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
@@ -28,36 +26,30 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @AllArgsConstructor
 @Builder
 @Entity
-@SQLDelete(sql = "UPDATE community_comment_like_dislike SET is_deleted = true WHERE community_comment_id = ? AND member_id = ?")
+@SQLDelete(sql = "UPDATE community_scrap SET is_deleted = true WHERE community_id = ? AND member_id = ?")
 @Where(clause = "is_deleted = false")
 @EntityListeners(AuditingEntityListener.class)
-public class CommunityCommentLikeDislike {
+public class CommunityScrap {
 
     @EmbeddedId
-    private CommunityCommentLikeDislikeId id;
+    private CommunityScrapId id;
 
     @MapsId("memberId")
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", insertable = false, updatable = false)
     private Member member;
 
-    @MapsId("commentId")
+    @MapsId("communityId")
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "community_comment_id", insertable = false, updatable = false)
-    private CommunityComment communityComment;
+    @JoinColumn(name = "community_id", insertable = false, updatable = false)
+    private Community community;
 
-    @Enumerated(EnumType.STRING)
-    private LikeStatus status;
+
+    private boolean isDeleted;
 
     @CreatedDate
     private LocalDateTime createdAt;
 
     @LastModifiedDate
     private LocalDateTime updatedAt;
-
-    private boolean isDeleted;
-
-    public void changeStatus(LikeStatus status) {
-        this.status = status;
-    }
 }
