+++ api-module/src/main/java/org/devridge/api/domain/community/entity/LikeStatus.java
@@ -1,0 +1,6 @@
+package org.devridge.api.domain.community.entity;
+
+public enum LikeStatus {
+    G,
+    B
+}
