+++ api-module/src/main/java/org/devridge/api/domain/community/entity/id/CommunityScrapId.java
@@ -1,4 +1,4 @@
-package org.devridge.api.domain.communityscrap;
+package org.devridge.api.domain.community.entity.id;
 
 import java.io.Serializable;
 import javax.persistence.Embeddable;
