+++ api-module/src/main/java/org/devridge/api/domain/community/repository/CommunityCommentLikeDislikeRepository.java
@@ -1,5 +1,7 @@
-package org.devridge.api.domain.communitycommentlikedislike;
+package org.devridge.api.domain.community.repository;
 
+import org.devridge.api.domain.community.entity.CommunityCommentLikeDislike;
+import org.devridge.api.domain.community.entity.id.CommunityCommentLikeDislikeId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
