+++ api-module/src/main/java/org/devridge/api/domain/community/repository/CommunityCommentRepository.java
@@ -1,6 +1,7 @@
-package org.devridge.api.domain.communitycomment;
+package org.devridge.api.domain.community.repository;
 
 import java.util.List;
+import org.devridge.api.domain.community.entity.CommunityComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
