+++ api-module/src/main/java/org/devridge/api/domain/community/repository/CommunityScrapRepository.java
@@ -1,5 +1,7 @@
-package org.devridge.api.domain.communityscrap;
+package org.devridge.api.domain.community.repository;
 
+import org.devridge.api.domain.community.entity.CommunityScrap;
+import org.devridge.api.domain.community.entity.id.CommunityScrapId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
