--- api-module/src/main/java/org/devridge/api/domain/communitycommentlikedislike/LikeStatus.java
@@ -1,6 +1,0 @@
-package org.devridge.api.domain.communitycommentlikedislike;
-
-public enum LikeStatus {
-    G,
-    B
-}
